refactor: move request structured logger to o11y
