+++ observability/request-logger.go
@@ -1,4 +1,4 @@
-package logger
+package observability
 
 import (
 	"fmt"
